# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Restaurante API",
    "version": "1.0.0",
    "description": "API REST para gestión de pedidos, reservas, ventas, inventario y clientes de restaurante.",
    "contact": { "email": "soporte@mirestaurante.com" }
  },
  "servers": [
    { "url": "http://localhost:9000/api", "description": "Desarrollo local" },
    { "url": "https://tu-dominio.com/api", "description": "Producción" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Venta": {
        "type": "object",
        "properties": {
          "ticket_id":  { "type": "string", "example": "TKT-20260420-000001" },
          "fecha":      { "type": "string", "format": "date" },
          "total":      { "type": "number" },
          "subtotal":   { "type": "number" },
          "tax_rate":   { "type": "number" },
          "tax_amount": { "type": "number" },
          "metodo_pago":{ "type": "string" },
          "cajero":     { "type": "string" },
          "items":      { "type": "array", "items": { "type": "object" } }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id":             { "type": "integer" },
          "nombre":         { "type": "string" },
          "email":          { "type": "string" },
          "rol":            { "type": "string", "enum": ["admin","gerente","staff","chef","mozo"] },
          "activo":         { "type": "boolean" },
          "restaurante_id": { "type": "integer" },
          "created_at":     { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/ventas": {
      "get": {
        "tags": ["Ventas"],
        "summary": "Listar ventas del día",
        "parameters": [
          { "name": "fecha", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": { "description": "Lista de ventas", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Venta" } } } } }
        }
      },
      "post": {
        "tags": ["Ventas"],
        "summary": "Registrar venta directa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items", "total", "metodo_pago"],
                "properties": {
                  "items":       { "type": "array", "items": { "type": "object" } },
                  "total":       { "type": "number" },
                  "metodo_pago": { "type": "string", "enum": ["efectivo","tarjeta","transferencia","qr"] },
                  "fecha":       { "type": "string", "format": "date" },
                  "skipStock":   { "type": "boolean", "description": "No descontar stock" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Venta registrada con ticket", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Venta" } } } }
        }
      }
    },
...
    "/ventas/resumen": {
      "get": {
        "tags": ["Ventas"],
        "summary": "Resumen por período",
        "parameters": [
          { "name": "periodo", "in": "query", "schema": { "type": "string", "enum": ["dia","semana","mes"] } }
        ],
        "responses": {
          "200": { "description": "Resumen de ventas" }
        }
      }
    },
    "/ventas/{id}": {
      "delete": {
        "tags": ["Ventas"],
        "summary": "Eliminar venta (admin)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "admin_code": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Venta eliminada" } }
      }
    },
...
    "/usuarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar usuarios del restaurante",
        "description": "Requiere rol admin o gerente.",
        "responses": {
          "200": { "description": "Lista de usuarios (sin password_hash)", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Usuario" } } } } }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear usuario",
        "description": "Requiere rol admin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "email", "password"],
                "properties": {
                  "nombre":   { "type": "string" },
                  "email":    { "type": "string" },
                  "password": { "type": "string", "minLength": 6 },
                  "rol":      { "type": "string", "enum": ["admin","gerente","staff","chef","mozo"], "default": "staff" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario creado" },
          "409": { "description": "Email ya registrado" }
        }
      }
    },
    "/usuarios/{id}": {
      "put": {
        "tags": ["Usuarios"],
        "summary": "Actualizar usuario",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre":   { "type": "string" },
                  "rol":      { "type": "string" },
                  "activo":   { "type": "boolean" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Usuario actualizado" } }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario (admin)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "admin_code": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Usuario eliminado" } }
      }
    },
...
  },
...
}